requirements.txt
asgiref==3.8.1
coverage==7.6.12
Django==5.1.6
django-rest-framework==0.1.0
djangorestframework==3.15.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.3
